+++ CHANGES.md
@@ -29,6 +29,7 @@ Python language
   * added `by` and `so` connectives in predicates
   * fixed overloading of `+` in Python code
   * list concatenation in logic with `add_list`
+  * recursive `#@function` not yet supported
 
 Provers
   * support for CVC5 1.0.0 (released April 6, 2022)